INS-309: move artifact manager interface to core/ledger
